Refactor refactor of HomePartners component
