+++ src/domain/home/partners/partnerLogoList.module.scss
@@ -1,0 +1,63 @@
+@import "layout";
+@import "variables";
+.big {
+    display: grid;
+    @include respond-below(md) {
+        justify-content: center;
+        gap: 2em;
+    }
+    @include respond-above(sm) {
+        grid-template-columns: repeat(2, 1fr);
+        gap: $xlargeMargin;
+    }
+    margin-bottom: $largeMargin;
+    .icon {
+        width: 2* $xlargeMargin;
+        height: $xlargeMargin;
+    
+        @include respond-above(sm) {
+            &:first-child {
+                justify-self: end;
+            }
+    
+            &:last-child {
+                justify-self: start;
+            }
+        }
+        img {
+            width: 100%;
+            height: auto;
+        }
+    
+    }    
+}
+.small {
+    display: grid;
+    gap: 2rem;
+    grid-template-columns: repeat(4, 1fr);
+    @include respond-above(sm) {
+        grid-template-columns: repeat(4, 1fr);
+    }
+    @include respond-above(md) {
+        grid-template-columns: repeat(6, 1fr);
+    }
+    .icon {
+        width: 75%;
+        height: auto;
+        justify-self: center;
+    
+        img {
+            width: 100%;
+            height: auto;
+        }
+        @include respond-above(sm) {
+            width: $xlargeMargin;
+            height: 0.67 * $xlargeMargin;
+        }
+        @include respond-above(md) {
+            width: 1.5 * $xlargeMargin;
+            height: $xlargeMargin;
+        }
+    
+    }
+}
