+++ src/domain/home/partners/types/partner.ts
@@ -1,0 +1,4 @@
+export interface Partner {
+  name: string;
+  icon: string;
+}
